Move util to @storybook/component
