+++ lib/ui/src/components/preview/tools/eject.tsx
@@ -1,9 +1,8 @@
 import global from 'global';
 import React from 'react';
-import { IconButton, Icons } from '@storybook/components';
+import { getStoryHref, IconButton, Icons } from '@storybook/components';
 import { Consumer, Combo } from '@storybook/api';
 import { Addon } from '@storybook/addons';
-import { getStoryHref } from '../utils/getStoryHref';
 
 const { PREVIEW_URL } = global;
